Extract Hal+Json format from API Root
And include it into API V3 as only that API is HAL-JSON only. Doing that we can create other APIs that want to use a different format.
